Move Invites V2 methods to their own namespace
